# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: security-atlas REST API
  version: v1
  description: |
    Operator-facing REST API for the security-atlas GRC platform.

    Two API surfaces:

    - **REST** (this document) — operator integration, dashboards, admin
      workflows. Authenticated via bearer tokens issued by
      `POST /v1/admin/credentials` (slice 034) or the
      `/auth/*` sign-in flow.
    - **gRPC** (specified separately in `proto/`) — Evidence SDK push
      (`evidence.IngestEvidence`), connector ingest, OSCAL bridge.

    The two surfaces stay separately specified; this spec covers only
    REST. See `Plans/EVIDENCE_SDK.md` for the dual-surface contract.

    Internal endpoints (`x-internal: true`) are filtered from the public
    Redoc render — they are operator probes (`/health`, `/metrics`,
    `/v1/version`, `/v1/install-state`) and not part of the consumer
    integration surface.
  license:
    name: license-pending
    url: https://github.com/mgoodric/security-atlas/blob/main/LICENSE
servers:
  - url: http://localhost:8080
    description: Local dev (default self-host)
  - url: https://your-deployment.example.com
    description: Self-hosted deployment (replace with your atlas host)
...
paths:
...
  /v1/me:
    get:
      summary: GET /v1/me
      tags: [me]
      operationId: get-v1-me
      security:
        - bearer: []
      responses:
        default:
          description: Default response envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ack"
    patch:
      summary: PATCH /v1/me
      tags: [me]
      operationId: patch-v1-me
      security:
        - bearer: []
      responses:
        default:
          description: Default response envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ack"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Any authenticated bearer credential issued by
        `POST /v1/admin/credentials` (slice 034) or established via the
        `/auth/*` OIDC / local sign-in flow. Token is opaque; the
        platform validates against the credstore or DB-backed apikeystore.
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Admin-only bearer credential. Same wire shape as `bearer`, but
        the platform rejects requests whose credential is not marked
        `is_admin=true` via the slice-035 OPA middleware AND the
        per-handler defense-in-depth role probe.
  schemas:
...
    Ack:
      type: object
      description: |
        Empty acknowledgment envelope. Used by mutation endpoints that
        do not return a representation of the mutated resource (e.g.
        `POST /v1/policies/{id}/publish`).
...